move config.py
