+++ app/core/security.py
@@ -3,7 +3,7 @@ from datetime import datetime, timezone, timedelta
 from typing import Optional
 from fastapi import Depends, HTTPException, status
 from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
-from core.config import JWT_ALGORITHM, JWT_EXPIRATION_HOURS, JWT_SECRET
+from config import JWT_ALGORITHM, JWT_EXPIRATION_HOURS, JWT_SECRET
 
 # Security
 security = HTTPBearer()
